+++ server/mongodb/actions/Volunteer.ts
@@ -8,12 +8,6 @@ import {
 } from "@/common_utils/types";
 import { PipelineStage } from "mongoose";
 import { flatten } from "mongo-dot-notation";
-import {
-  BlobServiceClient,
-  StorageSharedKeyCredential,
-  BlobDeleteOptions,
-  DeleteSnapshotsOptionType,
-} from "@azure/storage-blob";
 import User from "../models/User";
 import { deleteVerificationLogByEmail } from "./VerificationLog";
 
@@ -211,44 +205,4 @@ export const deleteVolunteer = async (email: string): Promise<null> => {
   await deleteVerificationLogByEmail(email);
 
   return null;
-};
-const deleteBlob = async (imageUrl: string) => {
-  const accountKey = process.env.AZURE_ACCOUNT_KEY;
-  const accountName = process.env.AZURE_ACCOUNT_NAME;
-  const containerName = process.env.AZURE_CONTAINER_NAME;
-  if (!accountKey || !accountName || !containerName) {
-    throw new Error("Azure account key or name is not defined");
-  }
-  const blobName = imageUrl.split("/").slice(4).join("/");
-  const sharedKeyCredential = new StorageSharedKeyCredential(
-    accountName,
-    accountKey,
-  );
-  const blobServiceClient = new BlobServiceClient(
-    `https://${accountName}.blob.core.windows.net`,
-    sharedKeyCredential,
-  );
-  const containerClient = blobServiceClient.getContainerClient(containerName);
-  const blobClient = containerClient.getBlockBlobClient(blobName);
-  const options: BlobDeleteOptions = {
-    deleteSnapshots: "include" as DeleteSnapshotsOptionType,
-  };
-  await blobClient.deleteIfExists(options);
-};
-
-export const postVolunteerImageLink = async (
-  email: string,
-  newImageLink: string,
-): Promise<IUser | null> => {
-  const user = await User.findOne<IUser>({ email });
-  if (user && user.imageLink) {
-    await deleteBlob(user.imageLink);
-  }
-
-  const updatedUser: IUser | null = await User.findOneAndUpdate(
-    { email },
-    { $set: { imageLink: newImageLink } },
-    { new: true },
-  );
-  return updatedUser;
 };
