+++ src/app/api/volunteer/profile-image/image-link/route.ts
@@ -1,5 +1,11 @@
 import APIWrapper from "@server/utils/APIWrapper";
-import { postVolunteerImageLink } from "@server/mongodb/actions/Volunteer";
+import { updateVolunteer } from "@server/mongodb/actions/Volunteer";
+import {
+  BlobDeleteOptions,
+  BlobServiceClient,
+  DeleteSnapshotsOptionType,
+  StorageSharedKeyCredential,
+} from "@azure/storage-blob";
 
 type PostRequest = {
   newImageLink: string;
@@ -9,8 +15,9 @@ type PostRequest = {
 export const POST = APIWrapper({
   config: {
     requireToken: true,
+    requireVolunteer: true,
   },
-  handler: async (req) => {
+  handler: async (req, currentUser) => {
     const requestData = (await req.json()) as PostRequest;
     const { email, newImageLink } = requestData;
     if (!requestData) {
@@ -22,8 +29,31 @@ export const POST = APIWrapper({
     if (!newImageLink) {
       throw new Error("Missing image link in request data");
     }
-    const user = await postVolunteerImageLink(email, newImageLink);
 
+    const accountKey = process.env.AZURE_ACCOUNT_KEY;
+    const accountName = process.env.AZURE_ACCOUNT_NAME;
+    const containerName = process.env.AZURE_CONTAINER_NAME;
+    if (!accountKey || !accountName || !containerName) {
+      throw new Error("Azure account key or name is not defined");
+    }
+
+    const blobName = currentUser!.imageLink.split("/").slice(4).join("/");
+    const sharedKeyCredential = new StorageSharedKeyCredential(
+      accountName,
+      accountKey,
+    );
+    const blobServiceClient = new BlobServiceClient(
+      `https://${accountName}.blob.core.windows.net`,
+      sharedKeyCredential,
+    );
+    const containerClient = blobServiceClient.getContainerClient(containerName);
+    const blobClient = containerClient.getBlockBlobClient(blobName);
+    const options: BlobDeleteOptions = {
+      deleteSnapshots: "include" as DeleteSnapshotsOptionType,
+    };
+    await blobClient.deleteIfExists(options);
+
+    const user = await updateVolunteer(email, { imageLink: newImageLink });
     return user;
   },
 });
